adapt patch of query column
